refactor: move WebSocketTransport into transport dir
