+++ ably/realtime/realtime_channel.py
@@ -4,7 +4,7 @@ import logging
 from typing import Optional
 
 from ably.realtime.connection import ConnectionState
-from ably.realtime.websockettransport import ProtocolMessageAction
+from ably.transport.websockettransport import ProtocolMessageAction
 from ably.rest.channel import Channel
 from ably.types.message import Message
 from ably.util.eventemitter import EventEmitter
