+++ test/ably/realtimechannel_test.py
@@ -1,7 +1,7 @@
 import asyncio
 import pytest
 from ably.realtime.realtime_channel import ChannelState
-from ably.realtime.websockettransport import ProtocolMessageAction
+from ably.transport.websockettransport import ProtocolMessageAction
 from ably.types.message import Message
 from test.ably.restsetup import RestSetup
 from test.ably.utils import BaseAsyncTestCase, random_string
